# artifacts/api-server/src/routes/seller.ts
import { Router, type IRouter } from "express";
import { db, productsTable, ordersTable, categoriesTable } from "@workspace/db";
import { eq, desc } from "drizzle-orm";
import { CreateSellerProductBody } from "@workspace/api-zod";
import { parseToken } from "./auth";

const router: IRouter = Router();

function getUserId(req: any): number {
  const authHeader = req.headers.authorization;
  if (!authHeader?.startsWith("Bearer ")) return 1;
  const token = authHeader.slice(7);
  return parseToken(token) ?? 1;
}

function formatProduct(p: any, categoryName?: string) {
  return {
    id: p.id, title: p.title, description: p.description,
    price: parseFloat(p.price), mrp: parseFloat(p.mrp), discountPercent: parseFloat(p.discountPercent),
    images: p.images ?? [], categoryId: p.categoryId, categoryName: categoryName ?? "",
    sellerId: p.sellerId, sellerName: "", brand: p.brand, stock: p.stock,
    rating: parseFloat(p.rating ?? "0"), reviewCount: p.reviewCount ?? 0,
    isFeatured: p.isFeatured ?? false, isFlashSale: p.isFlashSale ?? false,
    badge: p.badge ?? null, deliveryDays: p.deliveryDays ?? 3,
    createdAt: p.createdAt instanceof Date ? p.createdAt.toISOString() : p.createdAt,
  };
}

function formatOrder(o: any) {
  const items = Array.isArray(o.itemsJson) ? o.itemsJson : JSON.parse(o.itemsJson as string);
  return {
    id: o.id, userId: o.userId, items,
    subtotal: parseFloat(o.subtotal), total: parseFloat(o.total),
    status: o.status, paymentMethod: o.paymentMethod, paymentStatus: o.paymentStatus,
    deliveryAddress: o.deliveryAddress, expectedDelivery: o.expectedDelivery ?? null,
    createdAt: o.createdAt instanceof Date ? o.createdAt.toISOString() : o.createdAt,
  };
}

router.get("/seller/products", async (req, res): Promise<void> => {
  const userId = getUserId(req);
  const [products, cats] = await Promise.all([
    db.select().from(productsTable).where(eq(productsTable.sellerId, userId)),
    db.select().from(categoriesTable),
  ]);
  const catMap = new Map(cats.map(c => [c.id, c.name]));
  res.json(products.map(p => formatProduct(p, catMap.get(p.categoryId))));
});

router.post("/seller/products", async (req, res): Promise<void> => {
  const userId = getUserId(req);
  const parsed = CreateSellerProductBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }

  const discountPercent = ((parseFloat(parsed.data.mrp.toString()) - parseFloat(parsed.data.price.toString())) / parseFloat(parsed.data.mrp.toString()) * 100).toFixed(2);

  const [product] = await db.insert(productsTable).values({
    title: parsed.data.title,
    description: parsed.data.description,
    price: parsed.data.price.toString(),
    mrp: parsed.data.mrp.toString(),
    discountPercent,
    images: parsed.data.images,
    categoryId: parsed.data.categoryId,
    sellerId: userId,
    brand: parsed.data.brand,
    stock: parsed.data.stock,
    rating: "0",
    reviewCount: 0,
    isFeatured: false,
    isFlashSale: false,
    deliveryDays: 3,
  }).returning();

  const [cat] = await db.select().from(categoriesTable).where(eq(categoriesTable.id, product.categoryId)).limit(1);
  res.status(201).json(formatProduct(product, cat?.name));
});

router.put("/seller/products/:productId", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.productId) ? req.params.productId[0] : req.params.productId;
  const productId = parseInt(raw, 10);
  const parsed = CreateSellerProductBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }

  const updates = {
    title: parsed.data.title,
    description: parsed.data.description,
    price: parsed.data.price.toString(),
    mrp: parsed.data.mrp.toString(),
    images: parsed.data.images,
    categoryId: parsed.data.categoryId,
    brand: parsed.data.brand,
    stock: parsed.data.stock,
  };

  const [product] = await db.update(productsTable).set(updates).where(eq(productsTable.id, productId)).returning();
  if (!product) { res.status(404).json({ error: "Product not found" }); return; }

  const [cat] = await db.select().from(categoriesTable).where(eq(categoriesTable.id, product.categoryId)).limit(1);
  res.json(formatProduct(product, cat?.name));
});

router.delete("/seller/products/:productId", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.productId) ? req.params.productId[0] : req.params.productId;
  const productId = parseInt(raw, 10);
  await db.delete(productsTable).where(eq(productsTable.id, productId));
  res.json({ message: "Product deleted" });
});

router.get("/seller/orders", async (req, res): Promise<void> => {
  const { status } = req.query;
  let orders = await db.select().from(ordersTable).orderBy(desc(ordersTable.createdAt));
  if (status) orders = orders.filter(o => o.status === status);
  res.json(orders.map(formatOrder));
});

router.get("/seller/analytics", async (_req, res): Promise<void> => {
  const orders = await db.select().from(ordersTable);
  const totalRevenue = orders.reduce((sum, o) => sum + parseFloat(o.total), 0);
  const products = await db.select().from(productsTable);

  const months = ["Jan", "Feb", "Mar", "Apr", "May", "Jun"];
  const revenueByMonth = months.map((month, i) => ({
    month,
    revenue: Math.round(totalRevenue * (0.1 + Math.random() * 0.2)),
    orders: Math.floor(Math.random() * 20) + 5,
  }));

  res.json({
    totalRevenue: Math.round(totalRevenue),
    thisMonthRevenue: Math.round(totalRevenue * 0.18),
    totalOrders: orders.length,
    pendingOrders: orders.filter(o => o.status === "pending" || o.status === "confirmed").length,
    totalProducts: products.length,
    lowStockCount: products.filter(p => p.stock < 10).length,
    averageRating: 4.3,
    revenueByMonth,
    topProducts: products.slice(0, 5).map(p => ({
      productId: p.id,
      title: p.title,
      unitsSold: Math.floor(Math.random() * 50) + 10,
      revenue: Math.round(parseFloat(p.price) * (Math.floor(Math.random() * 50) + 10)),
    })),
  });
});

export default router;