# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZenKart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wishlist/{productId}:
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Added to wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove product from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # REVIEWS
...
  /seller/products:
    get:
      operationId: listSellerProducts
      tags: [seller]
      summary: List seller's products
      responses:
        "200":
          description: Seller products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createSellerProduct
      tags: [seller]
      summary: Create a new product listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /seller/products/{productId}:
    put:
      operationId: updateSellerProduct
      tags: [seller]
      summary: Update a seller product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteSellerProduct
      tags: [seller]
      summary: Delete a seller product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /seller/orders:
    get:
      operationId: listSellerOrders
      tags: [seller]
      summary: List orders for seller
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Seller orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /wallet/add-money:
    post:
      operationId: addMoneyToWallet
      tags: [wallet]
      summary: Add money to wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMoneyBody"
      responses:
        "200":
          description: Money added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"

  # NOTIFICATIONS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        mrp:
          type: number
        discountPercent:
          type: number
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        brand:
          type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        isFlashSale:
          type: boolean
        badge:
          type: string
          nullable: true
        deliveryDays:
          type: integer
        createdAt:
          type: string
      required: [id, title, price, mrp, discountPercent, images, categoryId, sellerId, brand, stock, rating, reviewCount, isFeatured, isFlashSale, createdAt]
...
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productTitle:
          type: string
        productImage:
          type: string
        quantity:
          type: integer
        price:
          type: number
        variant:
          type: string
          nullable: true
      required: [productId, productTitle, productImage, quantity, price]
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, returned]
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        deliveryAddress:
          type: string
        expectedDelivery:
          type: string
        createdAt:
          type: string
      required: [id, userId, items, subtotal, total, status, paymentMethod, paymentStatus, deliveryAddress, createdAt]
...
    WishlistItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        addedAt:
          type: string
      required: [id, productId, product, addedAt]
...
    CreateProductBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        mrp:
          type: number
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        brand:
          type: string
        stock:
          type: integer
      required: [title, description, price, mrp, images, categoryId, brand, stock]
...
    Wallet:
      type: object
      properties:
        balance:
          type: number
        zenCoins:
          type: integer
        zenCoinsValue:
          type: number
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"
      required: [balance, zenCoins, zenCoinsValue, recentTransactions]
    WalletTransaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [credit, debit, cashback, refund, commission]
        amount:
          type: number
        description:
          type: string
        reference:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, type, amount, description, createdAt]
    AddMoneyBody:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [upi, card, netbanking]
      required: [amount, paymentMethod]
...